#1318 helper function for scala and java collection converter
